+++ docs/changelog.md
@@ -2,11 +2,27 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+## [0.35.0] -- 2022-08-25
+
+### Changed 
+
+- Optimized converting Projects to and from dict. Now, `to_dict(extended=True)` returns only essential properties to save space and time.
+- Small refactors.
+
+### Fixed
+
+- Initialization of `peppy.Project` from `pandas.DataFrame`. Now `from_pandas()` can receive sample table, subsample table and config file
+- Multiple bugs introduced during initialization of the project with custom index column names
+
+### Added
+- New test cases and test data
+
 ## [0.34.0] -- 2022-08-17
 
 ### Changed 
 
-Way of initialization project from dictionary. Now it's possible as follows: Project().from_dict()
+- Way of initialization project from dictionary. Now it's possible as follows: `Project().from_dict()`
+- 
 ### Fixed
 
 - Fix error that was raised when duplicated sample in `sample_table` had different read types (single-end mixed with paired-end).